FuncGenPlugin: Add decorators
Decorators for functions are in the list of possible extensions to tc39/proposal-decorators (see https://github.com/tc39/proposal-decorators/blob/master/EXTENSIONS.md#function-decorators-and-annotations). One of possible workarounds is the use of decorators in a function expression. The decorators added in this commit are to be merged after the workaround solution is agreed on, implemented and tested.